+++ Solana/src/main/java/com/solana/models/buffer/BufferInfo.kt
@@ -1,0 +1,12 @@
+package com.solana.models.buffer
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+class BufferInfo<T>(
+    var data: Buffer<T>? = null,
+    val executable: Boolean,
+    val lamports: Double,
+    val owner: String?,
+    val rentEpoch: Double
+)
